Update swift tests

+++ scripts/setup_quickstart_framework.sh
@@ -27,12 +27,18 @@ mkdir -p Firebase/
 if [[ ! -z "$NON_FIREBASE_SDKS" ]]; then
   REPO="${REPO}" NON_FIREBASE_SDKS="${NON_FIREBASE_SDKS}" "${REPO}"/scripts/build_non_firebase_sdks.sh
 fi
-mv "${HOME}"/ios_frameworks/Firebase/Firebase.h Firebase/
-mv "${HOME}"/ios_frameworks/Firebase/module.modulemap Firebase/
+if [ ! -f "Firebase/Firebase.h" ]; then
+  mv "${HOME}"/ios_frameworks/Firebase/Firebase.h Firebase/
+fi
+if [ ! -f "Firebase/module.modulemap" ]; then
+  mv "${HOME}"/ios_frameworks/Firebase/module.modulemap Firebase/
+fi
 for file in "$@"
 do
+  if [ ! -f "Firebase/${file}" ]; then
   # Not override framework if a framework with the same name was moved early..
-  mv -n ${file} Firebase/
+    mv -n ${file} Firebase/
+	fi
 done
 
 ../scripts/add_framework_script.rb  "${SAMPLE}" "${TARGET}" Firebase
